Refactor auth config
