+++ src/server/auth/server/saml_test.go
@@ -3,8 +3,6 @@ package server
 import (
 	"bytes"
 	"encoding/base64"
-	"encoding/xml"
-	"errors"
 	"fmt"
 	"net/http"
 	"net/url"
@@ -12,12 +10,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/crewjam/saml"
-	"github.com/gogo/protobuf/proto"
-
 	"github.com/pachyderm/pachyderm/src/client/auth"
 	"github.com/pachyderm/pachyderm/src/client/pkg/require"
-	"github.com/pachyderm/pachyderm/src/server/pkg/backoff"
 	tu "github.com/pachyderm/pachyderm/src/server/pkg/testutil"
 )
 
@@ -52,571 +46,9 @@ func SimpleSAMLIDPMetadata(t *testing.T) []byte {
 		</EntityDescriptor>`)
 }
 
-// requireConfigsEqual compares 'expected' and 'actual' using 'proto.Equal', but
-// also formats any XML they contain to remove e.g. whitespace discrepancies
-// that we want to ignore
-func requireConfigsEqual(t testing.TB, expected, actual *auth.AuthConfig) {
-	e, a := proto.Clone(expected).(*auth.AuthConfig), proto.Clone(actual).(*auth.AuthConfig)
-	// Format IdP metadata (which is serialized XML) to fix e.g. whitespace
-	formatIDP := func(c *auth.IDProvider) {
-		if c.SAML == nil || c.SAML.MetadataXML == nil {
-			return
-		}
-		var d saml.EntityDescriptor
-		require.NoError(t, xml.Unmarshal(c.SAML.MetadataXML, &d))
-		c.SAML.MetadataXML = tu.MustMarshalXML(t, &d)
-	}
-	for _, idp := range e.IDProviders {
-		formatIDP(idp)
-	}
-	for _, idp := range a.IDProviders {
-		formatIDP(idp)
-	}
-	require.True(t, proto.Equal(e, a),
-		"expected: %v\n           actual: %v", e, a)
-}
-
-// TestSetGetConfigBasic sets an auth config and then retrieves it, to make
-// sure it's stored propertly
-func TestSetGetConfigBasic(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-	deleteAll(t)
-
-	adminClient := getPachClient(t, admin)
-
-	// Set a configuration
-	conf := &auth.AuthConfig{
-		LiveConfigVersion: 0,
-		IDProviders: []*auth.IDProvider{{
-			Name:        "idp",
-			Description: "fake IdP for testing",
-			SAML: &auth.IDProvider_SAMLOptions{
-				MetadataXML: SimpleSAMLIDPMetadata(t),
-			},
-		}},
-		SAMLServiceOptions: &auth.AuthConfig_SAMLServiceOptions{
-			ACSURL: "http://acs", MetadataURL: "http://metadata",
-		},
-	}
-	_, err := adminClient.SetConfiguration(adminClient.Ctx(),
-		&auth.SetConfigurationRequest{Configuration: conf})
-	require.NoError(t, err)
-
-	// Read the configuration that was just written
-	configResp, err := adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	conf.LiveConfigVersion = 1 // increment version
-	requireConfigsEqual(t, conf, configResp.Configuration)
-}
-
-// TestGetSetConfigAdminOnly confirms that only cluster admins can get/set the
-// auth config
-func TestGetSetConfigAdminOnly(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-	deleteAll(t)
-
-	alice := tu.UniqueString("alice")
-	anonClient, aliceClient, adminClient := getPachClient(t, ""), getPachClient(t, alice), getPachClient(t, admin)
-
-	// Confirm that the auth config starts out empty
-	configResp, err := adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	requireConfigsEqual(t, &auth.AuthConfig{}, configResp.Configuration)
-
-	// Alice tries to set the current configuration and fails
-	conf := &auth.AuthConfig{
-		LiveConfigVersion: 0,
-		IDProviders: []*auth.IDProvider{{
-			Name:        "idp",
-			Description: "fake IdP for testing",
-			SAML: &auth.IDProvider_SAMLOptions{
-				MetadataXML: SimpleSAMLIDPMetadata(t),
-			},
-		}},
-		SAMLServiceOptions: &auth.AuthConfig_SAMLServiceOptions{
-			ACSURL: "http://acs", MetadataURL: "http://metadata",
-		},
-	}
-	_, err = aliceClient.SetConfiguration(aliceClient.Ctx(),
-		&auth.SetConfigurationRequest{
-			Configuration: conf,
-		})
-	require.YesError(t, err)
-	require.Matches(t, "not authorized", err.Error())
-	require.Matches(t, "admin", err.Error())
-	require.Matches(t, "SetConfiguration", err.Error())
-
-	// Confirm that alice didn't modify the configuration by retrieving the empty
-	// config
-	configResp, err = adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	requireConfigsEqual(t, &auth.AuthConfig{}, configResp.Configuration)
-
-	// Modify the configuration and make sure anon can't read it, but alice and
-	// admin can
-	_, err = adminClient.SetConfiguration(adminClient.Ctx(),
-		&auth.SetConfigurationRequest{
-			Configuration: conf,
-		})
-	require.NoError(t, err)
-
-	// Confirm that anon can't read the config
-	_, err = anonClient.GetConfiguration(anonClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.YesError(t, err)
-	require.Matches(t, "no authentication token", err.Error())
-
-	// Confirm that alice and admin can read the config
-	configResp, err = aliceClient.GetConfiguration(aliceClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	conf.LiveConfigVersion = 1 // increment version
-	requireConfigsEqual(t, conf, configResp.Configuration)
-
-	configResp, err = adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	requireConfigsEqual(t, conf, configResp.Configuration)
-}
-
-// TestRMWConfigConflict does two conflicting R+M+W operation on a config
-// and confirms that one operation fails
-func TestRMWConfigConflict(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-	deleteAll(t)
-
-	adminClient := getPachClient(t, admin)
-
-	// Set a configuration
-	conf := &auth.AuthConfig{
-		LiveConfigVersion: 0,
-		IDProviders: []*auth.IDProvider{{
-			Name:        "idp",
-			Description: "fake IdP for testing",
-			SAML: &auth.IDProvider_SAMLOptions{
-				MetadataXML: SimpleSAMLIDPMetadata(t),
-			},
-		}},
-		SAMLServiceOptions: &auth.AuthConfig_SAMLServiceOptions{
-			ACSURL: "http://acs", MetadataURL: "http://metadata",
-		},
-	}
-
-	// Set an initial configuration
-	_, err := adminClient.SetConfiguration(adminClient.Ctx(),
-		&auth.SetConfigurationRequest{Configuration: conf})
-	require.NoError(t, err)
-
-	// Read the configuration that was just written
-	configResp, err := adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	conf.LiveConfigVersion = 1 // increment version
-	requireConfigsEqual(t, conf, configResp.Configuration)
-
-	// modify the config twice
-	mod2 := proto.Clone(configResp.Configuration).(*auth.AuthConfig)
-	mod2.IDProviders = []*auth.IDProvider{{
-		Name: "idp_2", Description: "fake IDP for testing, #2",
-		SAML: &auth.IDProvider_SAMLOptions{
-			MetadataXML: SimpleSAMLIDPMetadata(t),
-		}}}
-	mod3 := proto.Clone(configResp.Configuration).(*auth.AuthConfig)
-	mod3.IDProviders = []*auth.IDProvider{{
-		Name: "idp_3", Description: "fake IDP for testing, #3",
-		SAML: &auth.IDProvider_SAMLOptions{
-			MetadataXML: SimpleSAMLIDPMetadata(t),
-		}}}
-
-	// Apply both changes -- the second should fail
-	_, err = adminClient.SetConfiguration(adminClient.Ctx(), &auth.SetConfigurationRequest{
-		Configuration: mod2,
-	})
-	require.NoError(t, err)
-	_, err = adminClient.SetConfiguration(adminClient.Ctx(), &auth.SetConfigurationRequest{
-		Configuration: mod3,
-	})
-	require.YesError(t, err)
-	require.Matches(t, "config version", err.Error())
-
-	// Read the configuration and make sure that only #2 was applied
-	configResp, err = adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	mod2.LiveConfigVersion = 2 // increment version
-	requireConfigsEqual(t, mod2, configResp.Configuration)
-}
-
-// TestSetNilConfig sets a config, then overwrites it with a nil config to
-// disable the SAML service (this tests that an empty config validates and
-// disables the SAML service)
-func TestSetNilConfig(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-	deleteAll(t)
-
-	adminClient := getPachClient(t, admin)
-
-	// Set a configuration
-	conf := &auth.AuthConfig{
-		LiveConfigVersion: 0,
-		IDProviders: []*auth.IDProvider{{
-			Name:        "idp",
-			Description: "fake IdP for testing",
-			SAML: &auth.IDProvider_SAMLOptions{
-				MetadataXML: SimpleSAMLIDPMetadata(t)},
-		}},
-		SAMLServiceOptions: &auth.AuthConfig_SAMLServiceOptions{
-			ACSURL: "http://acs", MetadataURL: "http://metadata",
-		},
-	}
-	_, err := adminClient.SetConfiguration(adminClient.Ctx(),
-		&auth.SetConfigurationRequest{Configuration: conf})
-	require.NoError(t, err)
-
-	// Read the configuration that was just written
-	configResp, err := adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	conf.LiveConfigVersion = 1 // increment version
-	requireConfigsEqual(t, conf, configResp.Configuration)
-
-	// Set the nil config
-	_, err = adminClient.SetConfiguration(adminClient.Ctx(),
-		&auth.SetConfigurationRequest{})
-	require.NoError(t, err)
-
-	// Get the empty config
-	configResp, err = adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	requireConfigsEqual(t, &auth.AuthConfig{}, configResp.Configuration)
-
-	// TODO Make sure SAML has been deactivated
-}
-
-// TestSetEmptyConfig is like TestSetNilConfig, but it passes a non-nil but
-// empty config to disable SAML
-func TestSetEmptyConfig(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-	deleteAll(t)
-
-	adminClient := getPachClient(t, admin)
-
-	// Set a configuration
-	conf := &auth.AuthConfig{
-		LiveConfigVersion: 0,
-		IDProviders: []*auth.IDProvider{{
-			Name:        "idp",
-			Description: "fake IdP for testing",
-			SAML: &auth.IDProvider_SAMLOptions{
-				MetadataXML: SimpleSAMLIDPMetadata(t)},
-		}},
-		SAMLServiceOptions: &auth.AuthConfig_SAMLServiceOptions{
-			ACSURL: "http://acs", MetadataURL: "http://metadata",
-		},
-	}
-	_, err := adminClient.SetConfiguration(adminClient.Ctx(),
-		&auth.SetConfigurationRequest{Configuration: conf})
-	require.NoError(t, err)
-
-	// Read the configuration that was just written
-	configResp, err := adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	conf.LiveConfigVersion = 1 // increment version
-	requireConfigsEqual(t, conf, configResp.Configuration)
-
-	// Set the empty config
-	_, err = adminClient.SetConfiguration(adminClient.Ctx(),
-		&auth.SetConfigurationRequest{
-			Configuration: &auth.AuthConfig{},
-		})
-	require.NoError(t, err)
-
-	// Get the empty config
-	configResp, err = adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	requireConfigsEqual(t, &auth.AuthConfig{}, configResp.Configuration)
-
-	// TODO Make sure SAML has been deactivated
-}
-
-// TestGetEmptyConfig sets a config, then Deactivates+Reactivates auth, then
-// calls GetConfig on an empty cluster to be sure the config was cleared
-func TestGetEmptyConfig(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-	deleteAll(t)
-
-	adminClient := getPachClient(t, admin)
-
-	// Set a configuration
-	conf := &auth.AuthConfig{
-		LiveConfigVersion: 0,
-		IDProviders: []*auth.IDProvider{{
-			Name:        "idp",
-			Description: "fake IdP for testing",
-			SAML: &auth.IDProvider_SAMLOptions{
-				MetadataXML: SimpleSAMLIDPMetadata(t)},
-		}},
-		SAMLServiceOptions: &auth.AuthConfig_SAMLServiceOptions{
-			ACSURL: "http://acs", MetadataURL: "http://metadata",
-		},
-	}
-	_, err := adminClient.SetConfiguration(adminClient.Ctx(),
-		&auth.SetConfigurationRequest{Configuration: conf})
-	require.NoError(t, err)
-
-	// Read the configuration that was just written
-	configResp, err := adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	conf.LiveConfigVersion = 1 // increment version
-	requireConfigsEqual(t, conf, configResp.Configuration)
-
-	// Deactivate auth
-	_, err = adminClient.Deactivate(adminClient.Ctx(), &auth.DeactivateRequest{})
-	require.NoError(t, err)
-
-	// Wait for auth to be deactivated
-	require.NoError(t, backoff.Retry(func() error {
-		_, err := adminClient.WhoAmI(adminClient.Ctx(), &auth.WhoAmIRequest{})
-		if err != nil && auth.IsErrNotActivated(err) {
-			return nil // WhoAmI should fail when auth is deactivated
-		}
-		return errors.New("auth is not yet deactivated")
-	}, backoff.NewTestingBackOff()))
-
-	// Try to set and get the configuration, and confirm that the calls have been
-	// deactivated
-	_, err = adminClient.SetConfiguration(adminClient.Ctx(),
-		&auth.SetConfigurationRequest{Configuration: conf})
-	require.YesError(t, err)
-	require.Matches(t, "activated", err.Error())
-
-	_, err = adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.YesError(t, err)
-	require.Matches(t, "activated", err.Error())
-
-	// activate auth
-	activateResp, err := adminClient.Activate(adminClient.Ctx(), &auth.ActivateRequest{
-		GitHubToken: "admin",
-	})
-	require.NoError(t, err)
-	adminClient.SetAuthToken(activateResp.PachToken)
-
-	// Wait for auth to be re-activated
-	require.NoError(t, backoff.Retry(func() error {
-		_, err := adminClient.WhoAmI(adminClient.Ctx(), &auth.WhoAmIRequest{})
-		return err
-	}, backoff.NewTestingBackOff()))
-
-	// Try to get the configuration, and confirm that the config is now empty
-	configResp, err = adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	requireConfigsEqual(t, &auth.AuthConfig{}, configResp.Configuration)
-
-	// Set the configuration (again)
-	conf.LiveConfigVersion = 0 // reset to 0 (since config is empty)
-	_, err = adminClient.SetConfiguration(adminClient.Ctx(),
-		&auth.SetConfigurationRequest{Configuration: conf})
-	require.NoError(t, err)
-
-	// Get the configuration, and confirm that the config has been updated
-	configResp, err = adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	conf.LiveConfigVersion = 1 // increment version
-	requireConfigsEqual(t, conf, configResp.Configuration)
-}
-
-// TestValidateConfigErrNoName tests that SetConfig rejects configs with unnamed
-// ID providers
-func TestValidateConfigErrNoName(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-	deleteAll(t)
-	adminClient := getPachClient(t, admin)
-
-	conf := &auth.AuthConfig{
-		LiveConfigVersion: 0,
-		IDProviders: []*auth.IDProvider{&auth.IDProvider{
-			Description: "fake IdP for testing",
-			SAML: &auth.IDProvider_SAMLOptions{
-				MetadataXML: SimpleSAMLIDPMetadata(t)},
-		}},
-		SAMLServiceOptions: &auth.AuthConfig_SAMLServiceOptions{
-			ACSURL: "http://acs", MetadataURL: "http://metadata",
-		},
-	}
-	_, err := adminClient.SetConfiguration(adminClient.Ctx(),
-		&auth.SetConfigurationRequest{Configuration: conf})
-	require.YesError(t, err)
-	require.Matches(t, "must have a name", err.Error())
-
-	// Make sure config change wasn't applied
-	configResp, err := adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	requireConfigsEqual(t, &auth.AuthConfig{}, configResp.Configuration)
-}
-
-// TestValidateConfigErrReservedName tests that SetConfig rejects configs that
-// try to use a reserved name
-func TestValidateConfigErrReservedName(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-	deleteAll(t)
-	adminClient := getPachClient(t, admin)
-
-	for _, name := range []string{"github", "robot", "pipeline"} {
-		conf := &auth.AuthConfig{
-			LiveConfigVersion: 0,
-			IDProviders: []*auth.IDProvider{&auth.IDProvider{
-				Name:        name,
-				Description: "fake IdP for testing",
-				SAML: &auth.IDProvider_SAMLOptions{
-					MetadataXML: SimpleSAMLIDPMetadata(t)},
-			}},
-			SAMLServiceOptions: &auth.AuthConfig_SAMLServiceOptions{
-				ACSURL: "http://acs", MetadataURL: "http://metadata",
-			},
-		}
-		_, err := adminClient.SetConfiguration(adminClient.Ctx(),
-			&auth.SetConfigurationRequest{Configuration: conf})
-		require.YesError(t, err)
-		require.Matches(t, "reserved", err.Error())
-	}
-
-	// Make sure config change wasn't applied
-	configResp, err := adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	requireConfigsEqual(t, &auth.AuthConfig{}, configResp.Configuration)
-}
-
-// TestValidateConfigErrNoType tests that SetConfig rejects configs that
-// have an IDProvider (IDP) with no type
-func TestValidateConfigErrNoType(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-	deleteAll(t)
-	adminClient := getPachClient(t, admin)
-
-	conf := &auth.AuthConfig{
-		LiveConfigVersion: 0,
-		IDProviders: []*auth.IDProvider{&auth.IDProvider{
-			Name:        "idp",
-			Description: "fake IdP for testing",
-		}},
-		SAMLServiceOptions: &auth.AuthConfig_SAMLServiceOptions{
-			ACSURL: "http://acs", MetadataURL: "http://metadata",
-		},
-	}
-	_, err := adminClient.SetConfiguration(adminClient.Ctx(),
-		&auth.SetConfigurationRequest{Configuration: conf})
-	require.YesError(t, err)
-	require.Matches(t, "type", err.Error())
-
-	// Make sure config change wasn't applied
-	configResp, err := adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	requireConfigsEqual(t, &auth.AuthConfig{}, configResp.Configuration)
-}
-
-// TestValidateConfigErrInvalidIDPMetadata tests that SetConfig rejects configs
-// that have a SAML IDProvider with invalid XML in their IDPMetadata
-func TestValidateConfigErrInvalidIDPMetadata(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-	deleteAll(t)
-	adminClient := getPachClient(t, admin)
-
-	conf := &auth.AuthConfig{
-		LiveConfigVersion: 0,
-		IDProviders: []*auth.IDProvider{&auth.IDProvider{
-			Name:        "idp",
-			Description: "fake IdP for testing",
-			SAML: &auth.IDProvider_SAMLOptions{
-				MetadataXML: []byte("invalid XML"),
-			},
-		}},
-		SAMLServiceOptions: &auth.AuthConfig_SAMLServiceOptions{
-			ACSURL: "http://acs", MetadataURL: "http://metadata",
-		},
-	}
-	_, err := adminClient.SetConfiguration(adminClient.Ctx(),
-		&auth.SetConfigurationRequest{Configuration: conf})
-	require.YesError(t, err)
-	require.Matches(t, "could not unmarshal", err.Error())
-
-	// Make sure config change wasn't applied
-	configResp, err := adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	requireConfigsEqual(t, &auth.AuthConfig{}, configResp.Configuration)
-}
-
-// TestValidateConfigErrInvalidIDPMetadata tests that SetConfig rejects configs
-// that have a SAML IDProvider with an invalid metadata URL
-func TestValidateConfigErrInvalidMetadataURL(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-	deleteAll(t)
-	adminClient := getPachClient(t, admin)
-
-	conf := &auth.AuthConfig{
-		LiveConfigVersion: 0,
-		IDProviders: []*auth.IDProvider{&auth.IDProvider{
-			Name:        "idp",
-			Description: "fake IdP for testing",
-			SAML: &auth.IDProvider_SAMLOptions{
-				MetadataURL: "invalid URL",
-			},
-		}},
-		SAMLServiceOptions: &auth.AuthConfig_SAMLServiceOptions{
-			ACSURL: "http://acs", MetadataURL: "http://metadata",
-		},
-	}
-	_, err := adminClient.SetConfiguration(adminClient.Ctx(),
-		&auth.SetConfigurationRequest{Configuration: conf})
-	require.YesError(t, err)
-	require.Matches(t, "URL", err.Error())
-
-	// Make sure config change wasn't applied
-	configResp, err := adminClient.GetConfiguration(adminClient.Ctx(),
-		&auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	requireConfigsEqual(t, &auth.AuthConfig{}, configResp.Configuration)
-}
-
-// TestValidateConfigErrRedundantIDPMetadata tests that SetConfig rejects
-// configs with a SAML IDProvider that has both explicit metadata and a metadata
-// URL
-func TestValidateConfigErrRedundantIDPMetadata(t *testing.T) {
+// TestValidateConfigMultipleSAMLIdPs tests that SetConfig rejects
+// configs with a multiple SAML IDProviders
+func TestValidateConfigMultipleSAMLIdPs(t *testing.T) {
 	if testing.Short() {
 		t.Skip("Skipping integration tests in short mode")
 	}
@@ -825,83 +257,4 @@ func TestGroupsBasic(t *testing.T) {
 	require.NoError(t, aliceClient.GetFile(dataRepo, "master", "/data", 0, 0, &buf))
 	require.NoError(t, err)
 	require.Equal(t, "file contents", buf.String())
-}
-
-// TestConfigDeadlock tests that Pachyderm's SAML endpoint releases Pachyderm's
-// config mutex (a bug fix)
-func TestConfigDeadlock(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-	deleteAll(t)
-	adminClient := getPachClient(t, admin)
-
-	var (
-		pachdSAMLAddress = tu.GetACSAddress(t, adminClient.GetAddress())
-		pachdACSURL      = fmt.Sprintf("http://%s/saml/acs", pachdSAMLAddress)
-		pachdMetadataURL = fmt.Sprintf("http://%s/saml/metadata", pachdSAMLAddress)
-	)
-	testIDP := tu.NewTestIDP(t, pachdACSURL, pachdMetadataURL)
-
-	// Get current configuration, and then set a configuration (this may need to
-	// be retried if e.g. scraping metadata fails)
-	configResp, err := adminClient.GetConfiguration(adminClient.Ctx(), &auth.GetConfigurationRequest{})
-	require.NoError(t, err)
-	_, err = adminClient.SetConfiguration(adminClient.Ctx(), &auth.SetConfigurationRequest{
-		Configuration: &auth.AuthConfig{
-			LiveConfigVersion: configResp.Configuration.LiveConfigVersion,
-			IDProviders: []*auth.IDProvider{
-				{
-					Name:        "idp_1",
-					Description: "fake IdP for testing",
-					SAML: &auth.IDProvider_SAMLOptions{
-						MetadataXML:    testIDP.Metadata(),
-						GroupAttribute: "memberOf",
-					},
-				},
-			},
-			SAMLServiceOptions: &auth.AuthConfig_SAMLServiceOptions{
-				ACSURL:      pachdACSURL,
-				MetadataURL: pachdMetadataURL,
-			},
-		},
-	})
-	require.NoError(t, err)
-
-	// Send a SAMLResponse to Pachyderm
-	alice, group := tu.UniqueString("alice"), tu.UniqueString("group")
-	aliceClient := getPachClient(t, "") // empty string b/c want anon client
-	tu.AuthenticateWithSAMLResponse(t, aliceClient, testIDP.NewSAMLResponse(alice, group))
-	who, err := aliceClient.WhoAmI(aliceClient.Ctx(), &auth.WhoAmIRequest{})
-	require.NoError(t, err)
-	require.Equal(t, "idp_1:"+alice, who.Username)
-
-	// get/set Pachyderm's configuration again, to make sure the config mutex has
-	// been released
-	require.NoErrorWithinT(t, 30*time.Second, func() error {
-		configResp, err = adminClient.GetConfiguration(adminClient.Ctx(), &auth.GetConfigurationRequest{})
-		if err != nil {
-			return err
-		}
-		_, err = adminClient.SetConfiguration(adminClient.Ctx(), &auth.SetConfigurationRequest{
-			Configuration: &auth.AuthConfig{
-				LiveConfigVersion: configResp.Configuration.LiveConfigVersion,
-				IDProviders: []*auth.IDProvider{
-					{
-						Name:        "idp_2",
-						Description: "fake IdP for testing",
-						SAML: &auth.IDProvider_SAMLOptions{
-							MetadataXML:    testIDP.Metadata(),
-							GroupAttribute: "memberOf",
-						},
-					},
-				},
-				SAMLServiceOptions: &auth.AuthConfig_SAMLServiceOptions{
-					ACSURL:      pachdACSURL,
-					MetadataURL: pachdMetadataURL,
-				},
-			},
-		})
-		return err
-	})
 }
